UPBGE: Support float for object life.
Even if the life is still based on 1 = 1/50 second, now the float values are allowed
if the time need to be more precise than 50 fps.

+++ source/blender/makesdna/DNA_actuator_types.h
@@ -87,7 +87,7 @@ typedef struct bSoundActuator {
 } bSoundActuator;
 
 typedef struct bEditObjectActuator {
-	int time;
+	float time;
 	short type, flag;
 	struct Object *ob;
 	struct Mesh *me;
